add new control buttons

+++ show_status.php
@@ -1,5 +1,9 @@
 <?php
 session_start();
+include __DIR__ . '/templates/password_dialog.php';
+include __DIR__ . '/templates/timer_dialog.php';
+include __DIR__ . '/templates/open_dialog.php'; 
+
 $username = $_SESSION["username"];
 $sql = "SELECT * FROM users WHERE username = :username";
 $stmt = $pdo->prepare($sql);
@@ -20,13 +24,13 @@ $protectionLevelPassword = $row['protection_level_password'];
 $openTime    = $row['open_time'];
 $lockedSince = $row['locked_since'];
 $timeLeft    = $row['time_left'];
-$closeButtonText = "CLOSE";
-
-// DEBUG
-// $lockStatus = 0;
-// $appLoggedIn = 1;
-// $appActive = 1;
-// $boxName = 191919;
+ 
+// DEBUG -> coomment out autorefresh in header.php
+/*  $lockStatus = 1;
+ $appLoggedIn = 1;
+ $appActive = 1;
+ $protectionLevelPassword = 1;
+ $boxName = 191919;  */
 
 if (!empty($openTime)) {
     $openTime = date("y-m-d H:i", strtotime($openTime));
@@ -34,14 +38,12 @@ if (!empty($openTime)) {
 
 echo '<div class="overlay-card">';
 
-// Hintergrundbild mit ID
+// Hintergrundbild - initialer Zustand aus PHP
 echo '<img id="bg-image" class="bg-image" alt="Background" ';
 if (($appLoggedIn==1) && ($boxName!=0) && ($appActive==1)) {
     if ($lockStatus == 0) {
-        $closeButtonText = "CLOSE";
         echo 'src="/assets/images/icon_box_open.png">';
-    } elseif ($lockStatus == 1) {
-        $closeButtonText = "OPEN";
+    } else {
         echo 'src="/assets/images/icon_box_closed.png">';
     }
 } else {
@@ -50,7 +52,7 @@ if (($appLoggedIn==1) && ($boxName!=0) && ($appActive==1)) {
 
 echo '<div class="card-content">';
 
-// Status-Nachricht mit ID
+// Status-Nachricht - initialer Zustand aus PHP
 $connectionStatusMessage = "";
 if ($appActive == 1) {
     if ($appLoggedIn == 1) {
@@ -67,19 +69,40 @@ if ($appActive == 1) {
 }
 echo '<div id="status-message" class="status-message">' . $connectionStatusMessage . '</div>';
 
-// Box-Control-Bereich mit ID
 echo '<div id="box-control-area">';
 
-// IMMER rendern, JS steuert die Sichtbarkeit
-echo '<div id="box-control-area">';
+// box-control-form - initialer Zustand aus PHP
+if (($appLoggedIn==1) && ($boxName!=0) && ($appActive==1)) {
+    echo '<div id="box-control-form">';
+} else {
     echo '<div id="box-control-form" style="display:none;">';
-        include __DIR__ . '/templates/box_control_form.php';
-    echo '</div>';
+}
+include __DIR__ . '/templates/box_control_form.php';
+echo '</div>';
+
+// locked-since - initialer Zustand aus PHP
+if (($appLoggedIn==1) && ($boxName!=0) && ($appActive==1) && ($lockStatus==1)) {
+    echo '<div id="locked-since" class="locked-since">' . $lockedSince . '<br></div>';
+} else {
     echo '<div id="locked-since" class="locked-since" style="display:none"></div>';
+}
+
+// time-left & open-time - initialer Zustand aus PHP
+if ($protectionLevelTimer==1) {
+    echo '<div id="time-left" class="time-left">' . $timeLeft . '</div>';
+    echo '<div id="open-time" class="protection-level-timer">' . $openTime . '</div>';
+} else {
     echo '<div id="time-left" class="time-left" style="display:none"></div>';
     echo '<div id="open-time" class="protection-level-timer" style="display:none"></div>';
+}
+
+// password-symbol - initialer Zustand aus PHP
+if ($protectionLevelPassword==1) {
+    echo '<div id="password-symbol" class="protection-level-password"><img class="password-symbol" src="/assets/images/lockme_symbol_password.png"></div>';
+} else {
     echo '<div id="password-symbol" class="protection-level-password" style="display:none"></div>';
- 
+}
+
 echo '</div>'; // box-control-area
 echo '</div>'; // card-content
 echo '</div>'; // overlay-card
